==> 07.compute-cosine-similarities/0.compute-cosine-similarity-crispr.ipynb <==
import pandas as pd
import utils
import numpy as np
from tqdm import tqdm
import logging

logging.basicConfig(format="%(levelname)s:%(asctime)s:%(name)s:%(message)s")
logging.getLogger("copairs").setLevel(logging.INFO)

operations = "wellpos_var_mad_int_featselect_harmony_PCA_corrected"
batch_size = 20000

### Prepare the data

#### Read the CRISPR parquet file

crispr_df = pd.read_parquet(
    f"../profiles/profiles_{operations}.parquet"
)

#### Remove empty wells and `negcon`s

crispr_df = utils.remove_negcon_empty_wells(crispr_df)
crispr_df.shape

#### Remove `poscon` wells.

crispr_df = crispr_df.query('Metadata_Symbol!="PLK1"').reset_index(drop=True)
crispr_df.shape

#### Remove ORF reagents without a phenotypic activity

crispr_phenotypic_activity_df = (
    pd.read_csv(
        f"../03.retrieve-annotations/output/phenotypic-activity-{operations}.csv.gz",
        usecols=["Metadata_JCP2022", "below_corrected_p"],
    )
    .rename(columns={"below_corrected_p": "Metadata_below_corrected_p"})
).query("Metadata_below_corrected_p==True")

crispr_phenotypic_activity_df.shape

crispr_df = (
    crispr_df.merge(crispr_phenotypic_activity_df, on="Metadata_JCP2022", how="inner")
    .reset_index(drop=True)
)

crispr_df.shape

### Create cosine similarity matrix

#### Add Approved Symbols

crispr_metadata = pd.read_csv(
    "../00.download-and-process-annotations/output/crispr.csv.gz",
    usecols=["Metadata_JCP2022", "Metadata_Approved_Symbol"],
)

crispr_df = crispr_df.drop(columns=["Metadata_Approved_symbol"]).merge(
    crispr_metadata, on="Metadata_JCP2022", how="left"
)

crispr_df = crispr_df.assign(
    Metadata_Gene_Symbol=lambda x: np.where(
        x.Metadata_Approved_Symbol.isna(), x.Metadata_Symbol, x.Metadata_Approved_Symbol
    )
)

crispr_df.shape

#### Create consensus profiles

consensus_df = utils.consensus(crispr_df, "Metadata_Gene_Symbol")
consensus_df.shape

pos_sameby = []
pos_diffby = ["Metadata_Gene_Symbol"]
neg_sameby = []
neg_diffby = []

metadata_df = utils.get_metadata(consensus_df)
feature_df = utils.get_featuredata(consensus_df)
feature_values = feature_df.values

cosine_sim_df = utils.cosine_similarity(
    metadata_df,
    feature_values,
    pos_sameby,
    pos_diffby,
    neg_sameby,
    neg_diffby,
    batch_size=batch_size,
)

cosine_sim_df.head()

cosine_sim_df.to_parquet(
    f"cos_sim/cosine_similarities_genes_crispr.parquet", index=True
)
