==> 07.compute-cosine-similarities/01.2.compute-cosine-similarity-percentile-crispr.ipynb <==
import pandas as pd
from scipy.stats import rankdata

similarity_df = pd.read_parquet("cos_sim/cosine_similarities_genes_crispr.parquet")

similarity_stacked = (
    similarity_df.stack()
    .reset_index()
    .rename(columns={0: "cosine_similarity", "level_0": "gene_1", "level_1": "gene_2"})
    .query("gene_1 != gene_2")
)

similarity_stacked["percentile"] = rankdata(similarity_stacked["cosine_similarity"], method="ordinal")/len(similarity_stacked)
similarity_stacked.head()

similarity_percentile_df = similarity_stacked.pivot(index="gene_1", columns="gene_2", values="percentile").fillna(0)
similarity_percentile_df.head()

similarity_percentile_df.to_parquet("cos_sim/cosine_similarities_percentile_genes_crispr.parquet")
