==> 07.compute-cosine-similarities/19.0.KG-sanity-check.ipynb <==
import pandas as pd

Quickly check the knowledge graph values for some pairs of genes.

# Add knowledge graph information and filter out any connections with low knowledge graph scores

knowledge_graph = pd.read_csv(
    "~/Downloads/orf_scores_merged.zip",
    sep="\t",
    usecols=["GENE1", "GENE2", "gene_mf__go", "gene_bp__go", "gene_pathway"],
).rename(
    columns={
        "GENE1": "gene_1",
        "GENE2": "gene_2",
        "gene_mf__go": "gene_mf",
        "gene_bp__go": "gene_bp",
    }
)

query = "gene_1 == 'CNN1' & gene_2 == 'CNN2' | gene_1 == 'CORO2B' & gene_2 == 'CORO2A' | gene_1 == 'VGLL3' & gene_2 == 'VGLL4'"

print(knowledge_graph.query(query).to_markdown(index=False))
