==> 07.compute-cosine-similarities/20.0.cluster-similarity-to-negcon-profiles.ipynb <==
import pandas as pd
import utils
import numpy as np

feature_group = "AreaShape"
feature_group_list = ["Texture", "Intensity", "RadialDistribution"]
compartments = ["Cells", "Cytoplasm", "Nuclei"]
channels = ["DNA", "RNA", "Mito", "ER", "AGP"]

orf_genesets = [
    ["SLC39A1", "ZBTB16"],
    ["SARS2", "ECH1", "DGUOK", "MRPS2", "LDHAL6B", "UQCRFS1"],
    ["MYT1", "LZTS2", "CHRM4", "GPR176", "TSC22D1"],
    ["SPDL1", "PAFAH1B1", "NDEL1", "NDE1", "HOOK2", "HOOK1"],
    ["MYT1", "INSYN1", "RNF41"],
    [
        "ZFP36L1",
        "PIK3R3",
        "NRBP1",
        "INSYN1",
        "HOXC8",
        "RAB40C",
        "RAB40B",
        "HTR1D",
        "TUBA3D",
    ],
]

crispr_genesets = [
    ["SLC39A1", "ZBTB16"],
    ["FOXO3", "TGFB1"],
    ["SARS2", "ECH1", "PVR", "LAIR1", "SLC1A5", "UQCRFS1"],
    ["MYT1", "CHRM4", "GPR176", "TSC22D1"],
    ["PIK3R3", "ZFP36L1", "HOXC8", "NRBP1", "RAB40B"],
]

orf_metadata_df = pd.read_csv(
    "../00.download-and-process-annotations/output/orf_metadata.tsv.gz",
    sep="\t",
    usecols=["Metadata_JCP2022", "Metadata_Symbol", "Metadata_pert_type"],
)

orf_profiles_with_feature_names_df = pd.read_parquet(
    "../profiles/profiles_wellpos_cc_var_mad_outlier_orf.parquet"
).merge(orf_metadata_df, on="Metadata_JCP2022", how="inner")

orf_profiles_with_feature_names_df.head()

crispr_metadata_df = pd.read_csv(
    "../00.download-and-process-annotations/output/crispr_metadata.tsv.gz",
    sep="\t",
    usecols=["Metadata_JCP2022", "Metadata_Symbol"],
)

crispr_controls_df = pd.DataFrame(
    {
        "Metadata_JCP2022": ["JCP2022_805264", "JCP2022_800001", "JCP2022_800002"],
        "Metadata_pert_type": ["poscon", "negcon", "negcon"],
    },
    index=[0, 1, 2],
)

crispr_profiles_with_feature_names_df = (
    pd.read_parquet("../profiles/profiles_wellpos_cc_var_mad_outlier_crispr.parquet")
    .merge(crispr_metadata_df, on="Metadata_JCP2022", how="inner")
    .merge(crispr_controls_df, on="Metadata_JCP2022", how="left")
    .fillna(value={"Metadata_pert_type": "trt"})
)

crispr_profiles_with_feature_names_df.head()

feature_names = utils.get_featurecols(orf_profiles_with_feature_names_df)

feature_name_interpretation_df = pd.DataFrame()

for feature_name in feature_names:
    df = pd.DataFrame(utils.parse_cp_features(feature_name), index=[0])
    feature_name_interpretation_df = pd.concat([feature_name_interpretation_df, df], ignore_index=True)

feature_name_interpretation_df.head()

for geneset in orf_genesets:
    compartment_feature_group_df = pd.DataFrame()
    for compartment in compartments:
        feature_cols = (
            feature_name_interpretation_df.query("feature_group==@feature_group")
            .query("compartment==@compartment")
            .feature.to_list()
        )

        profiles = orf_profiles_with_feature_names_df.query(
            "Metadata_Symbol in @geneset"
        )
        plates_with_genes = list(np.unique(profiles.Metadata_Plate.to_list()))
        negative_control_profiles = orf_profiles_with_feature_names_df.query(
            "Metadata_Plate in @plates_with_genes"
        ).query("Metadata_pert_type=='negcon'")

        consensus_profiles = utils.consensus(profiles, "Metadata_JCP2022")
        consensus_profiles = utils.consensus(consensus_profiles, "Metadata_pert_type")[
            feature_cols
        ].copy()
        consensus_negative_control_profiles = utils.consensus(
            negative_control_profiles, "Metadata_pert_type"
        )[feature_cols].copy()

        id = ["genes", "negcon"]

        feature_values = np.asarray(
            [
                np.abs(utils.get_featuredata(consensus_profiles)).values[0],
                np.abs(
                    utils.get_featuredata(consensus_negative_control_profiles)
                ).values[0],
            ]
        )

        cosine_sim = utils.cosine_similarity(id, feature_values).values[0, 1]

        df = pd.DataFrame(
            {
                "Compartment": compartment,
                "Feature_group": feature_group,
                "Similarity": cosine_sim,
            },
            index=[0],
        )

        compartment_feature_group_df = pd.concat(
            [compartment_feature_group_df, df], ignore_index=True
        )

    compartment_feature_group_df = compartment_feature_group_df.pivot(
        index="Feature_group", columns="Compartment", values="Similarity"
    )

    print('-'.join(geneset))
    print(compartment_feature_group_df.to_markdown())

for geneset in orf_genesets:
    channel_feature_group_df = pd.DataFrame()
    for channel in channels:
        for feature_group in feature_group_list:
            feature_cols = (
                feature_name_interpretation_df.query("feature_group==@feature_group")
                .query("channel==@channel")
                .feature.to_list()
            )

            profiles = orf_profiles_with_feature_names_df.query(
                "Metadata_Symbol in @geneset"
            )
            plates_with_genes = list(np.unique(profiles.Metadata_Plate.to_list()))
            negative_control_profiles = orf_profiles_with_feature_names_df.query(
                "Metadata_Plate in @plates_with_genes"
            ).query("Metadata_pert_type=='negcon'")

            consensus_profiles = utils.consensus(profiles, "Metadata_JCP2022")
            consensus_profiles = utils.consensus(consensus_profiles, "Metadata_pert_type")[
                feature_cols
            ].copy()
            consensus_negative_control_profiles = utils.consensus(
                negative_control_profiles, "Metadata_pert_type"
            )[feature_cols].copy()

            id = ["genes", "negcon"]

            feature_values = np.asarray(
                [
                    np.abs(utils.get_featuredata(consensus_profiles)).values[0],
                    np.abs(
                        utils.get_featuredata(consensus_negative_control_profiles)
                    ).values[0],
                ]
            )

            cosine_sim = utils.cosine_similarity(id, feature_values).values[0, 1]

            df = pd.DataFrame(
                {
                    "Channel": channel,
                    "Feature_group": feature_group,
                    "Similarity": cosine_sim,
                },
                index=[0],
            )

            channel_feature_group_df = pd.concat(
                [channel_feature_group_df, df], ignore_index=True
            )

    channel_feature_group_df = channel_feature_group_df.pivot(
        index="Feature_group", columns="Channel", values="Similarity"
    )

    print('-'.join(geneset))
    print(channel_feature_group_df.to_markdown())

feature_group = "AreaShape"
feature_group_list = ["Texture", "Intensity", "RadialDistribution"]
compartments = ["Cells", "Cytoplasm", "Nuclei"]
channels = ["DNA", "RNA", "Mito", "ER", "AGP"]

feature_names = utils.get_featurecols(crispr_profiles_with_feature_names_df)

feature_name_interpretation_df = pd.DataFrame()

for feature_name in feature_names:
    df = pd.DataFrame(utils.parse_cp_features(feature_name), index=[0])
    feature_name_interpretation_df = pd.concat([feature_name_interpretation_df, df], ignore_index=True)

feature_name_interpretation_df.head()

for geneset in crispr_genesets:
    compartment_feature_group_df = pd.DataFrame()
    for compartment in compartments:
        feature_cols = (
            feature_name_interpretation_df.query("feature_group==@feature_group")
            .query("compartment==@compartment")
            .feature.to_list()
        )

        profiles = crispr_profiles_with_feature_names_df.query(
            "Metadata_Symbol in @geneset"
        )
        plates_with_genes = list(np.unique(profiles.Metadata_Plate.to_list()))
        negative_control_profiles = crispr_profiles_with_feature_names_df.query(
            "Metadata_Plate in @plates_with_genes"
        ).query("Metadata_pert_type=='negcon'")

        consensus_profiles = utils.consensus(profiles, "Metadata_JCP2022")
        consensus_profiles = utils.consensus(consensus_profiles, "Metadata_pert_type")[
            feature_cols
        ].copy()
        consensus_negative_control_profiles = utils.consensus(
            negative_control_profiles, "Metadata_pert_type"
        )[feature_cols].copy()

        id = ["genes", "negcon"]

        feature_values = np.asarray(
            [
                np.abs(utils.get_featuredata(consensus_profiles)).values[0],
                np.abs(
                    utils.get_featuredata(consensus_negative_control_profiles)
                ).values[0],
            ]
        )

        cosine_sim = utils.cosine_similarity(id, feature_values).values[0, 1]

        df = pd.DataFrame(
            {
                "Compartment": compartment,
                "Feature_group": feature_group,
                "Similarity": cosine_sim,
            },
            index=[0],
        )

        compartment_feature_group_df = pd.concat(
            [compartment_feature_group_df, df], ignore_index=True
        )

    compartment_feature_group_df = compartment_feature_group_df.pivot(
        index="Feature_group", columns="Compartment", values="Similarity"
    )

    print('-'.join(geneset))
    print(compartment_feature_group_df.to_markdown())

for geneset in crispr_genesets:
    channel_feature_group_df = pd.DataFrame()
    for channel in channels:
        for feature_group in feature_group_list:
            feature_cols = (
                feature_name_interpretation_df.query("feature_group==@feature_group")
                .query("channel==@channel")
                .feature.to_list()
            )

            profiles = crispr_profiles_with_feature_names_df.query(
                "Metadata_Symbol in @geneset"
            )
            plates_with_genes = list(np.unique(profiles.Metadata_Plate.to_list()))
            negative_control_profiles = crispr_profiles_with_feature_names_df.query(
                "Metadata_Plate in @plates_with_genes"
            ).query("Metadata_pert_type=='negcon'")

            consensus_profiles = utils.consensus(profiles, "Metadata_JCP2022")
            consensus_profiles = utils.consensus(consensus_profiles, "Metadata_pert_type")[
                feature_cols
            ].copy()
            consensus_negative_control_profiles = utils.consensus(
                negative_control_profiles, "Metadata_pert_type"
            )[feature_cols].copy()

            id = ["genes", "negcon"]

            feature_values = np.asarray(
                [
                    np.abs(utils.get_featuredata(consensus_profiles)).values[0],
                    np.abs(
                        utils.get_featuredata(consensus_negative_control_profiles)
                    ).values[0],
                ]
            )

            cosine_sim = utils.cosine_similarity(id, feature_values).values[0, 1]

            df = pd.DataFrame(
                {
                    "Channel": channel,
                    "Feature_group": feature_group,
                    "Similarity": cosine_sim,
                },
                index=[0],
            )

            channel_feature_group_df = pd.concat(
                [channel_feature_group_df, df], ignore_index=True
            )

    channel_feature_group_df = channel_feature_group_df.pivot(
        index="Feature_group", columns="Channel", values="Similarity"
    )

    print('-'.join(geneset))
    print(channel_feature_group_df.to_markdown())
