==> 07.compute-cosine-similarities/23.0.checking-cosine-sim.ipynb <==
import pandas as pd

orf_genesets = [
    ["SLC39A1", "ZBTB16"],
    ["SARS2", "ECH1", "DGUOK", "MRPS2", "LDHAL6B", "UQCRFS1", "ACADVL", "HMGCL"],
    ["MYT1", "LZTS2", "CHRM4", "GPR176", "TSC22D1", "NEUROD1", "MSC"],
    ["SPDL1", "PAFAH1B1", "NDEL1", "NDE1", "HOOK2", "HOOK1"],
    ["MYT1", "INSYN1", "RNF41", "PAK4", "IKZF1"],
    [
        "ZFP36L1",
        "PIK3R3",
        "NRBP1",
        "INSYN1",
        "HOXC8",
        "RAB40C",
        "RAB40B",
        "HTR1D",
        "TUBA3D",
    ],
    [
        "YAP1",
        "WWTR1",
        "VGLL4",
        "PRKCE",
        "STK3",
        "CEP72",
        "IL20RB",
        "MTMR9",
        "TXNIP",
        "LDLRAD4",
    ],
]

crispr_genesets = [
    ["SLC39A1", "ZBTB16"],
    ["FOXO3", "TGFB1"],
    ["SARS2", "ECH1", "PVR", "LAIR1", "SLC1A5", "UQCRFS1", "CYP3A4", "F2"],
    ["MYT1", "CHRM4", "GPR176", "TSC22D1", "GHRHR", "SCX"],
    ["PIK3R3", "ZFP36L1", "HOXC8", "NRBP1", "RAB40B", "F10", "HTR1B"],
]

# Read ORF similarities

orf_similarities_df = pd.read_parquet("cos_sim/cosine_similarities_genes_orf.parquet")

# Read CRISPR similarities

crispr_similarities_df = pd.read_parquet("cos_sim/cosine_similarities_genes_crispr.parquet")

for gene in orf_genesets:
    print(orf_similarities_df.loc[gene, gene].to_markdown())
    print()

for gene in crispr_genesets:
    print(crispr_similarities_df.loc[gene, gene].to_markdown())
    print()

orf_similarity_rank = (
    orf_similarities_df.unstack()
    .reset_index()
    .rename(columns={"level_0": "gene_1", "level_1": "gene_2", 0: "cosine_similarity"})
    .query("gene_1 != gene_2")
    .sort_values(by="cosine_similarity", ascending=False)
    .reset_index(drop=True)
    .reset_index()
    .rename(columns={"index": "rank"})
    .sort_values(by="rank", ascending=False)
    .reset_index(drop=True)
    .reset_index()
    .rename(columns={"index": "reverse_rank"})
)

orf_similarity_rank.query("gene_1=='MYT1' and gene_2=='RNF41'")

crispr_similarities_df.loc["MYT1", "RNF41"]

crispr_similarity_rank = (
    crispr_similarities_df.unstack()
    .reset_index()
    .rename(columns={"level_0": "gene_1", "level_1": "gene_2", 0: "cosine_similarity"})
    .query("gene_1 != gene_2")
    .sort_values(by="cosine_similarity", ascending=False)
    .reset_index(drop=True)
    .reset_index()
    .rename(columns={"index": "rank"})
    .sort_values(by="rank", ascending=False)
    .reset_index(drop=True)
    .reset_index()
    .rename(columns={"index": "reverse_rank"})
)
